+++ R/install.r
@@ -43,6 +43,7 @@
 #' @param threads number of concurrent threads to use for installing
 #'   dependencies.
 #'   It defaults to the option \code{"Ncpus"} or \code{1} if unset.
+#' @param ... additional arguments passed to \code{\link{install.packages}}
 #' @export
 #' @family package installation
 #' @seealso \code{\link{with_debug}} to install packages with debugging flags
@@ -51,7 +52,8 @@ install <- function(pkg = ".", reload = TRUE, quick = FALSE, local = TRUE,
                     args = getOption("devtools.install.args"), quiet = FALSE,
                     dependencies = NA, build_vignettes = FALSE,
                     keep_source = getOption("keep.source.pkgs"),
-                    threads = getOption("Ncpus", 1)) {
+                    threads = getOption("Ncpus", 1),
+                    ...) {
 
   pkg <- as.package(pkg)
   check_build_tools(pkg)
@@ -62,10 +64,10 @@ install <- function(pkg = ".", reload = TRUE, quick = FALSE, local = TRUE,
   if (build_vignettes && missing(dependencies)) {
     dependencies <- TRUE
   }
-  install_deps(pkg, dependencies = dependencies, threads = threads)
+  install_deps(pkg, dependencies = dependencies, threads = threads, ...)
 
   # Build the package. Only build locally if it doesn't have vignettes
-  has_vignettes <- length(tools::pkgVignettes(dir = pkg$path)$doc > 0)
+  has_vignettes <- length(tools::pkgVignettes(dir = pkg$path)$docs > 0)
   if (local && !(has_vignettes && build_vignettes)) {
     built_path <- pkg$path
   } else {
@@ -98,7 +100,8 @@ install <- function(pkg = ".", reload = TRUE, quick = FALSE, local = TRUE,
 #' @examples
 #' \dontrun{install_deps(".")}
 install_deps <- function(pkg = ".", dependencies = NA,
-                         threads = getOption("Ncpus", 1)) {
+                         threads = getOption("Ncpus", 1),
+                         ...) {
   pkg <- as.package(pkg)
   info <- pkg_deps(pkg, dependencies)
 
@@ -116,7 +119,7 @@ install_deps <- function(pkg = ".", dependencies = NA,
 
   message("Installing dependencies for ", pkg$package, ":\n",
     paste(deps, collapse = ", "))
-  utils::install.packages(deps, dependencies = NA, Ncpus = threads)
+  utils::install.packages(deps, dependencies = NA, Ncpus = threads, ...)
   invisible(deps)
 }