+++ R/utils.r
@@ -51,6 +51,11 @@ read_dcf <- function(path) {
 }
 
 write_dcf <- function(path, desc) {
+  desc <- unlist(desc)
+  # Add back in continuation characters
+  desc <- gsub("\n[ \t]*\n", "\n .\n ", desc, perl = TRUE, useBytes = TRUE)
+  desc <- gsub("\n \\.([^\n])", "\n  .\\1", desc, perl = TRUE, useBytes = TRUE)
+
   text <- paste0(names(desc), ": ", desc, collapse = "\n")
 
   if (substr(text, nchar(text), 1) != "\n") {