+++ NEWS.md
@@ -1,3 +1,55 @@
+# devtools 1.7.1.9000
+
+* devtools now correctly writes out continuation lines in the `DESCRIPTION`.
+  (#709).
+
+* `package_deps()` determines all recursive dependencies and whether or not
+  they're up-to-date (#663).
+
+* `revdep_check()` creates directories if needed (#759).
+
+* `install_version()` now knows how to look in multiple repos (#721).
+
+* Add details for RTools 3.2 (#738)
+
+* `use_travis()` now suggests you link to the svg icon since that looks a 
+  little sharper.
+
+* `use_code_of_conduct()` adds a contributor code of conduct from 
+  http://contributor-covenant.org. (#729)
+
+* `use_test()` adds a new test file (#769, @krlmlr).
+
+* `system_check()` combines arguments with ` `, not `, `. (#753)
+
+* `install()` and `install_deps()` gain a `...` argument, so additional
+  arguments can be passed to `utils::install.packages()` (@jimhester, #712).
+
+* `lint()` gains a `cache` argument (@jimhester, #708).
+* `test()` gains an `...` argument so that additional arguments can be passed
+  to `testthat::test_dir` (@jimhester, #747)
+
+* export functions `RCMD()` and `system_check()` so they can be used by other 
+  packages. (@jimhester, #699).
+
+* `dev_help()` now uses `normalizePath()`. Hopefully this will make it more
+  likely to work if you're on windows and have a space in the path.
+
+* `NOT_CRAN` is no longer set automatically if it has been set externally to
+  allow overriding.
+
+* `CRAN: http://cran.rstudio.com/` is explicitly set in `travis.yml`, enabling cyclic dependency check
+
+* `install_svn` now optionally accepts revision number (#739, @lev-kuznetsov)
+
+* `release()` now shows file size before submitting to CRAN (#683, @krlmlr).
+
+* Previously, `devtools` ran all external R processes with `R --vanilla`.
+  Now it only suppresses user profiles, and constructs a custom `.Rprofile` to
+  override the default.  Currently, this `.Rprofile` sets up the `repos` option.
+  Among others, this enables the cyclic dependency check in `devtools::release`
+  (#602, @krlmlr).
+
 # devtools 1.7.0
 
 ## Improve reverse dependency checking
@@ -75,6 +127,9 @@ I've also tweaked the output of `revdep_maintainers()` so it's easier to copy an
 * `missing_s3()` works once again (#672)
 
 * Fixed scoping issues with `unzip()`.
+
+* `load_code()` now executes the package's code with the package's root as
+  working directory, just like `R CMD build` et al. (#640, @krlmlr).
 
 # devtools 1.6.1